+++ src/frontend/apps/impress/src/hooks/useClipboard.tsx
@@ -2,7 +2,6 @@ import {
   VariantType,
   useToastProvider,
 } from '@gouvfr-lasuite/cunningham-react';
-import { announce } from '@react-aria/live-announcer';
 import { useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
 
@@ -19,14 +18,12 @@ export const useClipboard = () => {
           toast(message, VariantType.SUCCESS, {
             duration: 3000,
           });
-          announce(message, 'polite');
         })
         .catch(() => {
           const message = errorMessage ?? t('Failed to copy to clipboard');
           toast(message, VariantType.ERROR, {
             duration: 3000,
           });
-          announce(message, 'assertive');
         });
     },
     [t, toast],